split addlist.rs
