+++ src/lib/mod.rs
@@ -1,4 +1,4 @@
-pub mod addlist;
+pub mod aggregate;
 pub mod config;
 pub mod errors;
 pub mod thread;
